sched: Introduce Per-Cluster DVFS Headroom
Typically, UI-demanding tasks are handled by little and big cores,
while the prime core is more power-hungry and leads to unnecessary boosts, causing power wastage.
From my observations, boosting the little cores provides the most significant performance improvement for UI tasks,
followed by a modest boost to the big cores, resulting in fewer stutters across the entire system.
To avoid excessive boosting, it ensures that when the utilization is equal to maximum capacity value or higher, util is returned as-is.

[Jprimero15: Adapt for big.LITTLE]